added cron conditions

+++ src/schedule/schedule.service.ts
@@ -3,6 +3,7 @@ import { Cron, CronExpression } from '@nestjs/schedule';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Campaign } from 'src/campaigns/entities/campaign.entity';
 import { Coupon } from 'src/coupons/entities/coupon.entity';
+import * as os from 'os';
 import { GateWayLog } from 'src/gateway-logs/entities/log.entity';
 import { Log } from 'src/logs/entities/log.entity';
 import {
@@ -45,45 +46,63 @@ export class ScheduleService {
    */
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async handleCron() {
-    // Set 'today' to midnight (00:00:00) to match campaigns ending today
-    const today = new Date();
-    today.setHours(0, 0, 0, 0);
-    console.log('Running campaign expiry check...');
+    const hostName = os.hostname();
+    const localUrl = 'http://localhost:3000';
+    if (
+      process.env.PROD_SERVER_HOST_NAME === hostName ||
+      process.env.DEV_SERVER_HOST_NAME === hostName ||
+      process.env.UAT_SERVER_HOST_NAME === hostName ||
+      process.env.LOCAL_SERVER_HOST_NAME === localUrl
+    ) {
+      // Set 'today' to midnight (00:00:00) to match campaigns ending today
+      const today = new Date();
+      today.setHours(0, 0, 0, 0);
+      console.log('Running campaign expiry check...');
 
-    const qb = this.campaignRepository
-      .createQueryBuilder('campaign')
-      .leftJoinAndSelect('campaign.locales', 'locale')
-      .where('campaign.end_date = :today', { today })
-      .andWhere('campaign.active = :active', { active: true });
-    // .andWhere('locale.language_code = :langCode', { langCode: 'en' });
+      const qb = this.campaignRepository
+        .createQueryBuilder('campaign')
+        .leftJoinAndSelect('campaign.locales', 'locale')
+        .where('campaign.end_date = :today', { today })
+        .andWhere('campaign.active = :active', { active: true });
+      // .andWhere('locale.language_code = :langCode', { langCode: 'en' });
 
-    const expiredCampaigns: any = await qb.getMany();
+      const expiredCampaigns: any = await qb.getMany();
 
-    for (const campaign of expiredCampaigns) {
-      campaign.active = false;
-      await this.campaignRepository.save(campaign);
-      console.log(`Deactivated campaign: ${campaign?.locales?.[0]?.name}`);
+      for (const campaign of expiredCampaigns) {
+        campaign.active = false;
+        await this.campaignRepository.save(campaign);
+        console.log(`Deactivated campaign: ${campaign?.locales?.[0]?.name}`);
+      }
     }
   }
 
   // Cron job to remove logs older than 30 days
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async removeOldLogs() {
-    const date = new Date();
-    date.setDate(date.getDate() - 30);
+    const hostName = os.hostname();
+    const localUrl = 'http://localhost:3000';
+    if (
+      process.env.PROD_SERVER_HOST_NAME === hostName ||
+      process.env.DEV_SERVER_HOST_NAME === hostName ||
+      process.env.UAT_SERVER_HOST_NAME === hostName ||
+      process.env.LOCAL_SERVER_HOST_NAME === localUrl
+    ) {
+      const date = new Date();
+      date.setDate(date.getDate() - 30);
 
-    try {
-      const result = await this.logRepository.delete({
-        createdAt: LessThan(date),
-      });
-      console.log(`Deleted ${result.affected} old logs.`);
+      try {
+        const result = await this.logRepository.delete({
+          createdAt: LessThan(date),
+        });
+        console.log(`Deleted ${result.affected} old logs.`);
 
-      const result1 = await this.gatewayLogRepository.delete({
-        createdAt: LessThan(date),
-      });
-      console.log(`Deleted ${result1.affected} old gateway logs.`);
-    } catch (error) {
-      console.error('Error deleting old logs:', error);
+        const result1 = await this.gatewayLogRepository.delete({
+          createdAt: LessThan(date),
+        });
+        console.log(`Deleted ${result1.affected} old gateway logs.`);
+      } catch (error) {
+        console.error('Error deleting old logs:', error);
+      }
     }
   }
 
@@ -106,162 +125,189 @@ export class ScheduleService {
 
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async unLockWalletPointsAndAddThemInAvailableBalance() {
-    console.log('Cron Unlock Wallet Points And AddThem In Available Balance');
-    // 1. Find all wallet transactions where unlock_date is today or earlier and status is 'pending'
-    // const today = new Date();
+    const hostName = os.hostname();
+    const localUrl = 'http://localhost:3000';
+    if (
+      process.env.PROD_SERVER_HOST_NAME === hostName ||
+      process.env.DEV_SERVER_HOST_NAME === hostName ||
+      process.env.UAT_SERVER_HOST_NAME === hostName ||
+      process.env.LOCAL_SERVER_HOST_NAME === localUrl
+    ) {
+      console.log('Cron Unlock Wallet Points And AddThem In Available Balance');
+      // 1. Find all wallet transactions where unlock_date is today or earlier and status is 'pending'
+      // const today = new Date();
 
-    // Find all transactions that should be unlocked
-    // To handle date-only unlock_date (e.g., '2025-08-12') vs. JS Date (with time),
-    // we need to find all transactions where unlock_date is <= today (date part only).
-    // We'll use a raw query or Between/LessThanOrEqual with date string.
-    const today = new Date();
-    today.setHours(0, 0, 0, 0); // Set to start of today
-    const transactionsToUnlock = await this.txRepo.find({
-      where: {
-        unlock_date: LessThanOrEqual(today),
-        status: WalletTransactionStatus.PENDING,
-        type: WalletTransactionType.EARN,
-      },
-      relations: ['wallet'],
-    });
-
-    // console.log('transactionsToUnlock', transactionsToUnlock, today);
-
-    for (const tx of transactionsToUnlock) {
-      const wallet = tx.wallet;
-      if (!wallet) continue;
-
-      // Move points from locked_balance to available_balance, so picking how much locked points are for this transaction
-      const point_balance = Number(tx.point_balance);
-      // Update wallet balances
-      wallet.locked_balance = Number(wallet.locked_balance) - point_balance;
-      wallet.available_balance =
-        Number(wallet.available_balance) + point_balance;
-      // Update transaction status to 'active'
-      tx.status = WalletTransactionStatus.ACTIVE;
-
-      // Save changes
-      await this.walletService.updateWalletBalances(wallet.id, {
-        ...wallet,
+      // Find all transactions that should be unlocked
+      // To handle date-only unlock_date (e.g., '2025-08-12') vs. JS Date (with time),
+      // we need to find all transactions where unlock_date is <= today (date part only).
+      // We'll use a raw query or Between/LessThanOrEqual with date string.
+      const today = new Date();
+      today.setHours(0, 0, 0, 0); // Set to start of today
+      const transactionsToUnlock = await this.txRepo.find({
+        where: {
+          unlock_date: LessThanOrEqual(today),
+          status: WalletTransactionStatus.PENDING,
+          type: WalletTransactionType.EARN,
+        },
+        relations: ['wallet'],
       });
-      await this.txRepo.save(tx);
+
+      // console.log('transactionsToUnlock', transactionsToUnlock, today);
+
+      for (const tx of transactionsToUnlock) {
+        const wallet = tx.wallet;
+        if (!wallet) continue;
+
+        // Move points from locked_balance to available_balance, so picking how much locked points are for this transaction
+        const point_balance = Number(tx.point_balance);
+        // Update wallet balances
+        wallet.locked_balance = Number(wallet.locked_balance) - point_balance;
+        wallet.available_balance =
+          Number(wallet.available_balance) + point_balance;
+        // Update transaction status to 'active'
+        tx.status = WalletTransactionStatus.ACTIVE;
+
+        // Save changes
+        await this.walletService.updateWalletBalances(wallet.id, {
+          ...wallet,
+        });
+        await this.txRepo.save(tx);
+      }
     }
   }
 
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async deleteExpiredWalletPoints() {
-    console.log('Running wallet points expiry check...');
+    const hostName = os.hostname();
+    const localUrl = 'http://localhost:3000';
+    if (
+      process.env.PROD_SERVER_HOST_NAME === hostName ||
+      process.env.DEV_SERVER_HOST_NAME === hostName ||
+      process.env.UAT_SERVER_HOST_NAME === hostName ||
+      process.env.LOCAL_SERVER_HOST_NAME === localUrl
+    ) {
+      console.log('Running wallet points expiry check...');
 
-    const today = new Date(); // or however you get your date
-    today.setHours(0, 0, 0, 0);
+      const today = new Date(); // or however you get your date
+      today.setHours(0, 0, 0, 0);
 
-    // getting all transactions that are active and have expiry_date equal to today
-    // and type is 'earn'
-    const expiringTransactions = await this.txRepo.find({
-      where: {
-        expiry_date: LessThanOrEqual(today),
-        status: WalletTransactionStatus.ACTIVE,
-        type: In([
-          WalletTransactionType.EARN,
-          WalletTransactionType.ADJUSTMENT,
-        ]),
-      },
-      relations: ['wallet'],
-    });
-
-    for (const tx of expiringTransactions) {
-      const wallet = tx.wallet;
-      if (!wallet) continue;
-
-      // 1️⃣ Get all earned + adjustment points in the transaction's lifetime
-      const earnedTxs = await this.txRepo.find({
+      // getting all transactions that are active and have expiry_date equal to today
+      // and type is 'earn'
+      const expiringTransactions = await this.txRepo.find({
         where: {
-          id: Not(tx.id), // Exclude the current transaction
-          wallet: { id: wallet.id },
+          expiry_date: LessThanOrEqual(today),
+          status: WalletTransactionStatus.ACTIVE,
           type: In([
             WalletTransactionType.EARN,
             WalletTransactionType.ADJUSTMENT,
           ]),
-          created_at: Between(tx.created_at, tx.expiry_date),
-          status: WalletTransactionStatus.ACTIVE,
         },
+        relations: ['wallet'],
       });
 
-      const totalEarned = earnedTxs.reduce(
-        (sum, e) => sum + Number(e.amount),
-        0,
-      );
-      const availableBalance = Number(wallet.available_balance);
-      const remainingPastBalance = availableBalance - totalEarned;
-      const hasSpentSomething = remainingPastBalance > tx.point_balance;
+      for (const tx of expiringTransactions) {
+        const wallet = tx.wallet;
+        if (!wallet) continue;
 
-      if (hasSpentSomething) {
-        // 2️⃣ Calculate unused portion of THIS expiring transaction
-        let unusedPoints = remainingPastBalance - Number(tx.point_balance);
-        if (unusedPoints < 0) unusedPoints = 0;
+        // 1️⃣ Get all earned + adjustment points in the transaction's lifetime
+        const earnedTxs = await this.txRepo.find({
+          where: {
+            id: Not(tx.id), // Exclude the current transaction
+            wallet: { id: wallet.id },
+            type: In([
+              WalletTransactionType.EARN,
+              WalletTransactionType.ADJUSTMENT,
+            ]),
+            created_at: Between(tx.created_at, tx.expiry_date),
+            status: WalletTransactionStatus.ACTIVE,
+          },
+        });
 
-        // 3️⃣ Deduct unused portion via addTransaction
-        if (unusedPoints > 0) {
-          await this.walletService.addTransaction(
-            {
-              wallet_id: wallet.id,
-              business_unit_id: wallet.business_unit.id,
-              type: WalletTransactionType.EXPIRE,
-              amount: 0,
-              status: WalletTransactionStatus.ACTIVE,
-              description: `Expired ${unusedPoints} unused points from transaction ${tx.id}`,
-              source_type: 'system',
-              source_id: tx.id, // Link to the original transaction
-              prev_available_points: wallet.available_balance,
-              points_balance: unusedPoints,
-            },
-            null, // system user ID or dedicated service account
-            true, // callingFromGateway = true to skip permission checks
-          );
+        const totalEarned = earnedTxs.reduce(
+          (sum, e) => sum + Number(e.amount),
+          0,
+        );
+        const availableBalance = Number(wallet.available_balance);
+        const remainingPastBalance = availableBalance - totalEarned;
+        const hasSpentSomething = remainingPastBalance > tx.point_balance;
+
+        if (hasSpentSomething) {
+          // 2️⃣ Calculate unused portion of THIS expiring transaction
+          let unusedPoints = remainingPastBalance - Number(tx.point_balance);
+          if (unusedPoints < 0) unusedPoints = 0;
+
+          // 3️⃣ Deduct unused portion via addTransaction
+          if (unusedPoints > 0) {
+            await this.walletService.addTransaction(
+              {
+                wallet_id: wallet.id,
+                business_unit_id: wallet.business_unit.id,
+                type: WalletTransactionType.EXPIRE,
+                amount: 0,
+                status: WalletTransactionStatus.ACTIVE,
+                description: `Expired ${unusedPoints} unused points from transaction ${tx.id}`,
+                source_type: 'system',
+                source_id: tx.id, // Link to the original transaction
+                prev_available_points: wallet.available_balance,
+                points_balance: unusedPoints,
+              },
+              null, // system user ID or dedicated service account
+              true, // callingFromGateway = true to skip permission checks
+            );
+          }
+          console.log(`Expired ${unusedPoints} points for wallet ${wallet.id}`);
         }
-        console.log(`Expired ${unusedPoints} points for wallet ${wallet.id}`);
       }
     }
   }
 
   @Cron(CronExpression.EVERY_DAY_AT_MIDNIGHT)
   async walletPointsExpiryCron() {
-    console.log('Wallet expire points cron started');
+    const hostName = os.hostname();
+    const localUrl = 'http://localhost:3000';
+    if (
+      process.env.PROD_SERVER_HOST_NAME === hostName ||
+      process.env.DEV_SERVER_HOST_NAME === hostName ||
+      process.env.UAT_SERVER_HOST_NAME === hostName ||
+      process.env.LOCAL_SERVER_HOST_NAME === localUrl
+    ) {
+      console.log('Wallet expire points cron started');
 
-    // Get start of today (midnight) to compare expiry dates
-    const today = new Date();
-    today.setHours(0, 0, 0, 0);
+      // Get start of today (midnight) to compare expiry dates
+      const today = new Date();
+      today.setHours(0, 0, 0, 0);
 
-    // 1. Find all active EARN transactions whose points should expire today or earlier
-    const transactionsToExpire = await this.txRepo.find({
-      where: {
-        expiry_date: LessThanOrEqual(today),
-        status: WalletTransactionStatus.ACTIVE,
-        type: WalletTransactionType.EARN,
-      },
-      relations: ['wallet'],
-    });
+      // 1. Find all active EARN transactions whose points should expire today or earlier
+      const transactionsToExpire = await this.txRepo.find({
+        where: {
+          expiry_date: LessThanOrEqual(today),
+          status: WalletTransactionStatus.ACTIVE,
+          type: WalletTransactionType.EARN,
+        },
+        relations: ['wallet'],
+      });
 
-    // console.log('transactionsToExpire :::', transactionsToExpire);
+      // console.log('transactionsToExpire :::', transactionsToExpire);
 
-    // 2. Process each transaction
-    for (const tx of transactionsToExpire) {
-      const wallet = tx.wallet;
-      if (!wallet) continue; // Skip if wallet relation is missing
+      // 2. Process each transaction
+      for (const tx of transactionsToExpire) {
+        const wallet = tx.wallet;
+        if (!wallet) continue; // Skip if wallet relation is missing
 
-      const point_balance = Number(tx.point_balance);
+        const point_balance = Number(tx.point_balance);
 
-      // Move expired points from available to locked balance
-      wallet.available_balance =
-        Number(wallet.available_balance) - point_balance;
-      wallet.locked_balance = Number(wallet.locked_balance) + point_balance;
+        // Move expired points from available to locked balance
+        wallet.available_balance =
+          Number(wallet.available_balance) - point_balance;
+        wallet.locked_balance = Number(wallet.locked_balance) + point_balance;
 
-      // Mark transaction as expired
-      tx.status = WalletTransactionStatus.EXPIRED;
+        // Mark transaction as expired
+        tx.status = WalletTransactionStatus.EXPIRED;
 
-      // Save changes
-      await this.walletService.updateWalletBalances(wallet.id, { ...wallet });
-      await this.txRepo.save(tx);
+        // Save changes
+        await this.walletService.updateWalletBalances(wallet.id, { ...wallet });
+        await this.txRepo.save(tx);
+      }
     }
   }
 }
